+++ MapboxNavigation/NavigationView.swift
@@ -1,6 +1,7 @@
 import UIKit
 import MapboxDirections
 
+
 /**
  A view that represents the root view of the MapboxNavigation drop-in UI.